Schiff mit einem zufälligen Punkt
